+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2023-05-31
+
 ### Added
 - Processing of an attachment and date as field type
 - Compatibility with TYPO3 v12
@@ -64,7 +66,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial pre-release
 
-[Unreleased]: https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/brotkrueml/typo3-jobrouter-base/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.3.0...v2.0.0
 [1.3.0]: https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.0.0...v1.1.0